services/requirements.txt
fastapi>=0.110
uvicorn>=0.30
python-dotenv>=1.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.2.0
langchain-text-splitters>=0.2.0
langsmith>=0.1.0
pinecone-client>=3.0.0
psycopg2-binary>=2.9
sqlalchemy>=2.0
